+++ resources/views/templates/page/_partials/refresh-modal.blade.php
@@ -1,4 +1,4 @@
-<x-chief::dialog.modal id="refresh-modal" title="Je sessie is verlopen" size="xs" class="z-1000!">
+<x-chief::dialog.modal id="refresh-modal" title="Je sessie is verlopen" size="xs" class="z-200">
     <div class="prose prose-dark prose-spacing">
         <p>Deze pagina was te lang inactief. Klik op de knop hieronder om de pagina opnieuw te laden.</p>
     </div>
@@ -12,7 +12,7 @@
     </x-slot>
 </x-chief::dialog.modal>
 
-<x-chief::dialog.modal id="error-modal" title="Er is iets misgelopen" size="xs" class="z-1000!">
+<x-chief::dialog.modal id="error-modal" title="Er is iets misgelopen" size="xs" class="z-200">
     <div class="prose prose-dark prose-spacing">
         <p>
             Er is een onverwachte fout opgetreden. Klik op de knop hieronder om de pagina opnieuw te laden.