+++ src/Table/UI/views/columns/text.blade.php
@@ -2,7 +2,7 @@
     @if ($item->hasLink())
         <a
             href="{{ $item->getLink() }}"
-            title="{{ $item->getValue($getLocale()) }}"
+            title="{{ $item->getOriginalValue($getLocale()) }}"
             {!! $item->shouldOpenInNewTab() ? 'target="_blank" rel="noopener"' : '' !!}
             {!! $item->getCustomAttributesAsString() !!}
             class="text-grey-800 leading-5 hover:underline hover:underline-offset-2"
@@ -10,7 +10,11 @@
             {!! $item->getValue($getLocale()) !!}
         </a>
     @else
-        <span class="text-grey-500 leading-5" {!! $item->getCustomAttributesAsString() !!}>
+        <span
+            title="{{ $item->getOriginalValue($getLocale()) }}"
+            class="text-grey-500 leading-5"
+            {!! $item->getCustomAttributesAsString() !!}
+        >
             {!! $item->getValue($getLocale()) !!}
         </span>
     @endif